requirements.txt
streamlit
pillow
plotly
folium
streamlit-folium

